New Development: No             JIRA Issue: Yes (CAS-6109)         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs: runUnitTest.main(['test_sdlist'])   Put in Release Notes: No              Module(s): RowAccumulator, averaging codes            Description: Fixed a trivial bug of precision in getTime.

+++ src/RowAccumulator.cpp
@@ -262,7 +262,7 @@ Vector<Float> RowAccumulator::getSpectrum() const
 
 Double RowAccumulator::getTime() const
 {
-  return timeSum_/Float(max(n_));
+  return timeSum_/Double(max(n_));
 }
 
 Double RowAccumulator::getInterval() const
